fixes to grammar
Made editing fixes, grammar, capitalization, etc.
rephrased some sentences that were ambiguous (to me)

Also this statement is not really correct "Gaussian filtering is highly effective in removing Gaussian noise from the image.": the main goal of the Gaussian filter is to attenuate the influence of pixels that are further away from the central pixel, when compared to pixels that are immediately adjacent, and do this 'smoothly', i.e. us coefficient that fall off smoothy. The gaussian effect is in spatial domain. On the other hand, if you are talking about Gaussian noise, you are talking about the intensity domain. Different things... One of the nice features of Gaussian vs box filter is that b/c of abrupt falloff of the box filter coefficient in the spatial domain (the box filter is a step function) we will get Gibbs effects, therefore aliasing. (the Fourier transform of a step function filter is a sinc function, the ripples in the sinc are what causes the aliasing). By contrast the Fourier transform of a Gaussian is also a Gaussian, 
therefore no ripples, no Gibbs effects..

In sum I would just remove that sentence.


Btw, this tutorial is great! Thanks for  and making it available putting it together.  We are going to use it in our class at JHU and encourage students to submit back errata.