I believe that the sin value should be negative
I am in a class where we implemented HoughLines by hand, while the above works fine for normal retangular and reflective images, it didn't work well with things like a triangle or in the case of trying to recognize other objects that are lines. I know a lot of people from this class have been using this section of code for their line drawing, and it's an unnecessary pain (in my opinion) so I'd like to propose changing it, what do you think?