Fix variable assignment problem with face detection example.
This commit fixes an error with the face detection example in the Python
tutorials. It currently assigns the result of a `cv2.rectangle()` call
to the variable `img`, which is supposed to contain the
image. The `cv2.rectangle()` call does not return anything,
so `img` is null, so the subsequent assignment of `roi_color` fails.

This fixes the problem by eliminating the variable assignment.