Update py_thresholding.rst
There is some inconsistency in the threshold.jpg image with the behaviour of TRUNC and TOZERO_INV. This is due to the default rescaling behaviour of pyplot.imshow(). It is well known that this can be easily corrected by adding the optional arg of vmin and vmax. I did this minor correction to the rst file and adding some other things I personally find useful along the way I read it. Please, if someone does review it, consider these and correct them (because I kind of edit it in a markdown-file-way, not knowing exactly how to do rst editing.)