Fixed minor error (?) in first example
Initially the loop was, 
for rho,theta in lines[0]:
    a = np.cos(theta)
    b = np.sin(theta)
    x0 = a*rho
    y0 = b*rho
    x1 = int(x0 + 1000*(-b))
    y1 = int(y0 + 1000*(a))
    x2 = int(x0 - 1000*(-b))
    y2 = int(y0 - 1000*(a))

this is incorrect because for rho,theta in lines[0] returns rho, theta values for only one of the lines, whereas lines is an array of several rho,theta pairs. I have updated the loop so that it considers all pairs of rho, theta values.