cv2.findContours() updated in newer versions
cv2.findContours() no longer makes changes on the source image. (Newer Versions)
It takes in only 2 arguments instead of 3 as shown in the block of code. Upon executing the code, error will be thrown stating "only returns two objects, you're trying to unpack it into three."
Added a small note, how to deal with it if the user is using a newer version of OpenCV later than 3.2